Adding options and minor tweaks

+++ src/Bank.java
@@ -1,19 +1,19 @@
 @SuppressWarnings("unused")
 public class Bank {
-    private float money;
+    private double money;
     private String user;
-    private float amount;
+    private double amount;
 
     public Bank(String user){
         this.user = user;
         System.out.println("Welcome To The Banking App "+user);
     }
 
-    public void addMoney(float amount){
+    public void addMoney(double amount){
         money += amount;
     }
 
-    public void withdrawMoney(float amount){
+    public void withdrawMoney(double amount){
         if (amount > money) {
             System.out.println("Cannot Withdraw Balance Low : "+money);
         }
